@@ -4,6 +4,7 @@ package org.egov.pgr.web.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.response.ResponseInfo;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.pgr.service.PGRService;
 import org.egov.pgr.util.PGRConstants;
 import org.egov.pgr.util.ResponseInfoFactory;
@@ -13,6 +14,7 @@ import org.egov.pgr.web.models.RequestSearchCriteria;
 import org.egov.pgr.web.models.ServiceRequest;
 import org.egov.pgr.web.models.ServiceResponse;
 import org.egov.pgr.web.models.ServiceWrapper;
+import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,13 +24,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
+import static org.egov.pgr.util.PGRConstants.INVALID_TENANT_ID_ERR_CODE;
+
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
 
 @Controller
 @RequestMapping("/v2")
@@ -65,9 +69,14 @@ public class RequestsApiController{
                                                               @Valid @ModelAttribute RequestSearchCriteria criteria) {
     	
     	String tenantId = criteria.getTenantId();
-        List<ServiceWrapper> serviceWrappers = pgrService.search(requestInfoWrapper.getRequestInfo(), criteria);
-        Map<String,Integer> dynamicData = pgrService.getDynamicData(tenantId);
-        
+        List<ServiceWrapper> serviceWrappers = null;
+        Map<String,Integer> dynamicData = null;
+        try {
+            serviceWrappers = pgrService.search(requestInfoWrapper.getRequestInfo(), criteria);
+            dynamicData = pgrService.getDynamicData(tenantId);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID_ERR_CODE, e.getMessage());
+        }
         int complaintsResolved = dynamicData.get(PGRConstants.COMPLAINTS_RESOLVED);
 	    int averageResolutionTime = dynamicData.get(PGRConstants.AVERAGE_RESOLUTION_TIME);
 	    int complaintTypes = pgrService.getComplaintTypes();
@@ -81,7 +90,12 @@ public class RequestsApiController{
 
     @RequestMapping(value = "request/_plainsearch", method = RequestMethod.POST)
     public ResponseEntity<ServiceResponse> requestsPlainSearchPost(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper, @Valid @ModelAttribute RequestSearchCriteria requestSearchCriteria) {
-        List<ServiceWrapper> serviceWrappers = pgrService.plainSearch(requestInfoWrapper.getRequestInfo(), requestSearchCriteria);
+        List<ServiceWrapper> serviceWrappers = null;
+        try {
+            serviceWrappers = pgrService.plainSearch(requestInfoWrapper.getRequestInfo(), requestSearchCriteria);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID_ERR_CODE, e.getMessage());
+        }
         ResponseInfo responseInfo = responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true);
         ServiceResponse response = ServiceResponse.builder().responseInfo(responseInfo).serviceWrappers(serviceWrappers).build();
         return new ResponseEntity<>(response, HttpStatus.OK);
@@ -100,7 +114,12 @@ public class RequestsApiController{
     @RequestMapping(value="/request/_count", method = RequestMethod.POST)
     public ResponseEntity<CountResponse> requestsCountPost(@Valid @RequestBody RequestInfoWrapper requestInfoWrapper,
                                                            @Valid @ModelAttribute RequestSearchCriteria criteria) {
-        Integer count = pgrService.count(requestInfoWrapper.getRequestInfo(), criteria);
+        Integer count = null;
+        try {
+            count = pgrService.count(requestInfoWrapper.getRequestInfo(), criteria);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(INVALID_TENANT_ID_ERR_CODE, e.getMessage());
+        }
         ResponseInfo responseInfo = responseInfoFactory.createResponseInfoFromRequestInfo(requestInfoWrapper.getRequestInfo(), true);
         CountResponse response = CountResponse.builder().responseInfo(responseInfo).count(count).build();
         return new ResponseEntity<>(response, HttpStatus.OK);
