@@ -6,11 +6,11 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.hibernate.validator.constraints.SafeHtml;
+import org.egov.tracer.annotations.CustomSafeHtml;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,7 +18,7 @@ import java.util.List;
  * Boundary
  */
 @Validated
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
+@jakarta.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2020-07-15T11:35:33.568+05:30")
 
 @Getter
 @Setter
@@ -28,23 +28,23 @@ import java.util.List;
 public class Boundary   {
 
         @NotNull
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("code")
         private String code = null;
 
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("name")
         private String name = null;
 
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("label")
         private String label = null;
 
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("latitude")
         private String latitude = null;
 
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("longitude")
         private String longitude = null;
 
@@ -52,7 +52,7 @@ public class Boundary   {
         @Valid
         private List<Boundary> children = null;
 
-        @SafeHtml
+        @CustomSafeHtml
         @JsonProperty("materializedPath")
         private String materializedPath = null;
 
