@@ -10,8 +10,8 @@ import lombok.ToString;
 import org.egov.pgr.web.models.Status;
 import org.springframework.validation.annotation.Validated;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 @Validated
 @AllArgsConstructor
