@@ -9,7 +9,7 @@ import lombok.NonNull;
 import lombok.Setter;
 import org.egov.common.contract.request.RequestInfo;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 
 @Getter
