@@ -1,14 +1,14 @@
 package org.egov.pgr.web.controllers;
 
 import org.egov.pgr.TestConfiguration;
-import org.junit.Ignore;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -17,8 +17,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 /**
 * API tests for RequestsApiController
 */
-@Ignore
-@RunWith(SpringRunner.class)
+@Disabled
+@ExtendWith(MockitoExtension.class)
 @WebMvcTest(RequestsApiController.class)
 @Import(TestConfiguration.class)
 public class RequestsApiControllerTest {
