@@ -3,6 +3,7 @@ package org.egov.pgr;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.egov.common.utils.MultiStateInstanceUtil;
 import org.egov.tracer.config.TracerConfiguration;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringApplication;
@@ -15,7 +16,7 @@ import java.util.TimeZone;
 
 @SpringBootApplication
 @ComponentScan(basePackages = { "org.egov.pgr", "org.egov.pgr.web.controllers" , "org.egov.pgr.config"})
-@Import({ TracerConfiguration.class })
+@Import({ TracerConfiguration.class, MultiStateInstanceUtil.class })
 public class PGRApp {
 
         @Value("${app.timezone}")
