@@ -6,7 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.validator.constraints.SafeHtml;
+import org.egov.tracer.annotations.CustomSafeHtml;
 
 import java.util.Set;
 
@@ -16,7 +16,7 @@ import java.util.Set;
 @Builder
 public class RequestSearchCriteria {
 
-    @SafeHtml
+    @CustomSafeHtml
     @JsonProperty("tenantId")
     private String tenantId;
 
@@ -29,11 +29,11 @@ public class RequestSearchCriteria {
     @JsonProperty("applicationStatus")
     private Set<String> applicationStatus;
 
-    @SafeHtml
+    @CustomSafeHtml
     @JsonProperty("mobileNumber")
     private String mobileNumber;
 
-    @SafeHtml
+    @CustomSafeHtml
     @JsonProperty("serviceRequestId")
     private String serviceRequestId;
 
@@ -85,7 +85,7 @@ public class RequestSearchCriteria {
         serviceRequestId
     }
 
-    @SafeHtml
+    @CustomSafeHtml
     @JsonProperty("accountId")
     private String accountId;
 
